- Added a way to group fields that shoud be navigatable by the previous/next accessorybar buttons - Upgrade plugin according to the NativeScript Plugin Seed
