+++ .gitignore
@@ -1,6 +1,16 @@
+.vscode
 .idea
-npm-debug.log
+.DS_Store
 *.js
-demo/hooks/
-demo/platforms/
-demo/node_modules/
+*.js.map
+*.log
+package-lock.json
+src/*.d.ts
+!src/index.d.ts
+!src/references.d.ts
+demo/*.d.ts
+!demo/references.d.ts
+demo/platforms
+node_modules
+publish/src
+publish/package
