--- .npmignore
@@ -1,4 +1,0 @@
-.idea
-.gitignore
-*.gif
-demo/
