+++ demo/app/main-page.ts
@@ -1,5 +1,5 @@
-import * as observable from "data/observable";
-import * as pages from "ui/page";
+import * as observable from "tns-core-modules/data/observable";
+import * as pages from "tns-core-modules/ui/page";
 import { HelloWorldModel } from "./main-view-model";
 
 // Event handler for Page "loaded" event attached in main-page.xml
