+++ demo/app/main-page.xml
@@ -1,42 +1,68 @@
-<Page xmlns="http://schemas.nativescript.org/tns.xsd" loaded="pageLoaded">
+<Page xmlns="http://schemas.nativescript.org/tns.xsd" xmlns:IQKbMgr="nativescript-iqkeyboardmanager" loaded="pageLoaded">
   <ScrollView>
     <StackLayout>
 
-      <GridLayout rows="auto, auto, auto, auto, auto, auto, auto, auto, auto" columns="*, auto" horizontalAlignment="stretch">
-        <Label row="0" colSpan="2" text="Flip these default settings and try the effect with the textfields below" textWrap="true" class="title"/>
+      <iOS>
+        <GridLayout rows="auto, auto, auto, auto, auto, auto, auto, auto, auto" columns="*, auto" horizontalAlignment="stretch">
+          <Label row="0" colSpan="2" text="Flip these default settings and try the effect with the textfields below" textWrap="true" class="title"/>
 
-        <Label row="1" text="Enable IQKeyboard" class="{{ iqKeyboardEnabled ? '' : 'disabled' }}"/>
-        <Switch row="1" checked="{{ iqKeyboardEnabled }}" tap="{{ toggleIQKeyboard }}"/>
+          <Label row="1" text="Enable IQKeyboard" class="{{ iqKeyboardEnabled ? '' : 'disabled' }}"/>
+          <Switch row="1" checked="{{ iqKeyboardEnabled }}" tap="{{ toggleIQKeyboard }}"/>
 
-        <Label row="2" text="Enable IQKeyboard toolbar" class="{{ iqKeyboardToolbarEnabled ? '' : 'disabled' }}"/>
-        <Switch row="2" checked="{{ iqKeyboardToolbarEnabled }}" tap="{{ toggleIQKeyboardToolbar }}"/>
+          <Label row="2" text="Enable IQKeyboard toolbar" class="{{ iqKeyboardToolbarEnabled ? '' : 'disabled' }}"/>
+          <Switch row="2" checked="{{ iqKeyboardToolbarEnabled }}" tap="{{ toggleIQKeyboardToolbar }}"/>
 
-        <Label row="3" text="Tap outside keeps keyboard open" class="{{ keepKeyboardOpenOnTouchOutside ? '' : 'disabled' }}"/>
-        <Switch row="3" checked="{{ keepKeyboardOpenOnTouchOutside }}" tap="{{ toggleKeepKeyboardOpen }}"/>
+          <Label row="3" text="Tap outside keeps keyboard open" class="{{ keepKeyboardOpenOnTouchOutside ? '' : 'disabled' }}"/>
+          <Switch row="3" checked="{{ keepKeyboardOpenOnTouchOutside }}" tap="{{ toggleKeepKeyboardOpen }}"/>
 
-        <Label row="4" text="Show 'hint' in toolbar" class="{{ showHintInToolbar ? '' : 'disabled' }}"/>
-        <Switch row="4" checked="{{ showHintInToolbar }}" tap="{{ toggleShowHintInToolbar }}"/>
+          <Label row="4" text="Show 'hint' in toolbar" class="{{ showHintInToolbar ? '' : 'disabled' }}"/>
+          <Switch row="4" checked="{{ showHintInToolbar }}" tap="{{ toggleShowHintInToolbar }}"/>
 
-        <Label row="5" text="Change toolbar 'Done' button" class="{{ toggleDoneButtonTextChanged ? '' : 'disabled' }}"/>
-        <Switch row="5" checked="{{ toggleDoneButtonTextChanged }}" tap="{{ toggleDoneButtonText }}"/>
+          <Label row="5" text="Change toolbar 'Done' button" class="{{ toggleDoneButtonTextChanged ? '' : 'disabled' }}"/>
+          <Switch row="5" checked="{{ toggleDoneButtonTextChanged }}" tap="{{ toggleDoneButtonText }}"/>
 
-        <Label row="6" text="Dark keyboard appearance" class="{{ keyboardAppearanceDark ? '' : 'disabled' }}"/>
-        <Switch row="6" checked="{{ keyboardAppearanceDark }}" tap="{{ toggleKeyboardAppearance }}"/>
+          <Label row="6" text="Dark keyboard appearance" class="{{ keyboardAppearanceDark ? '' : 'disabled' }}"/>
+          <Switch row="6" checked="{{ keyboardAppearanceDark }}" tap="{{ toggleKeyboardAppearance }}"/>
 
-        <Label row="7" text="Increased keyboard distance" class="{{ increaseKeyboardDistanceFromTextField ? '' : 'disabled' }}"/>
-        <Switch row="7" checked="{{ increaseKeyboardDistanceFromTextField }}" tap="{{ toggleIncreaseKeyboardDistanceFromTextField }}"/>
-      </GridLayout>
+          <Label row="7" text="Increased keyboard distance" class="{{ increaseKeyboardDistanceFromTextField ? '' : 'disabled' }}"/>
+          <Switch row="7" checked="{{ increaseKeyboardDistanceFromTextField }}" tap="{{ toggleIncreaseKeyboardDistanceFromTextField }}"/>
+        </GridLayout>
+      </iOS>
 
-      <TextField hint="Email"/>
-      <TextField hint="Password"/>
-      <TextField hint="3"/>
-      <TextField hint="4"/>
-      <TextField hint="5"/>
-      <TextField hint="6"/>
-      <TextField hint="7"/>
-      <TextField hint="8"/>
-      <TextField hint="9"/>
-      <TextField hint="10"/>
+      <IQKbMgr:PreviousNextView>
+        <StackLayout>
+          <StackLayout>
+            <TextField hint="Email"/>
+          </StackLayout>
+          <StackLayout>
+            <TextField hint="Password"/>
+          </StackLayout>
+          <StackLayout>
+            <TextField hint="3"/>
+          </StackLayout>
+          <StackLayout>
+            <TextField hint="4"/>
+          </StackLayout>
+          <StackLayout>
+            <TextField hint="5"/>
+          </StackLayout>
+          <StackLayout>
+            <TextField hint="6"/>
+          </StackLayout>
+          <StackLayout>
+            <TextField hint="7"/>
+          </StackLayout>
+          <StackLayout>
+            <TextField hint="8"/>
+          </StackLayout>
+          <StackLayout>
+            <TextField hint="9"/>
+          </StackLayout>
+          <StackLayout>
+            <TextField hint="10"/>
+          </StackLayout>
+        </StackLayout>
+      </IQKbMgr:PreviousNextView>
     </StackLayout>
   </ScrollView>
 </Page>
