+++ demo/package.json
@@ -3,10 +3,22 @@
     "id": "org.nativescript.iqkeyboardmanagerdemo",
     "tns-ios": {
       "version": "3.4.0"
+    },
+    "tns-android": {
+      "version": "3.4.1"
     }
   },
   "dependencies": {
-    "nativescript-iqkeyboardmanager": "file:../../nativescript-IQKeyboardManager",
-    "tns-core-modules": "^3.4.0"
+    "nativescript-iqkeyboardmanager": "../src",
+    "tns-core-modules": "~3.4.0"
+  },
+  "devDependencies": {
+    "babel-traverse": "6.26.0",
+    "babel-types": "6.26.0",
+    "babylon": "6.18.0",
+    "lazy": "1.0.11",
+    "nativescript-dev-typescript": "^0.6.0",
+    "tns-platform-declarations": "~3.4.1",
+    "typescript": "~2.6.0"
   }
 }
