--- package.json
@@ -1,32 +1,0 @@
-{
-  "name": "nativescript-iqkeyboardmanager",
-  "version": "1.2.0",
-  "description": "NativeScript wrapper of the popular IQKeyboardManager iOS library",
-  "typings": "index.d.ts",
-  "nativescript": {
-    "platforms": {
-      "ios": "1.4.0"
-    }
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/tjvantoll/nativescript-IQKeyboardManager.git"
-  },
-  "keywords": [
-    "NativeScript"
-  ],
-  "author": "TJ VanToll <tj.vantoll@gmail.com> (http://tjvantoll.com/)",
-  "license": {
-    "type": "MIT"
-  },
-  "contributors": [
-    {
-      "name": "Eddy Verbruggen",
-      "email": "eddyverbruggen@gmail.com"
-    }
-  ],
-  "homepage": "https://github.com/tjvantoll/nativescript-IQKeyboardManager",
-  "bugs": {
-    "url": "https://github.com/tjvantoll/nativescript-IQKeyboardManager/issues"
-  }
-}
