+++ publish/package.json
@@ -1,0 +1,9 @@
+{
+  "name": "nativescript-publish",
+  "version": "1.0.0",
+  "description": "Publish helper",
+  "devDependencies": {
+    "ncp": "^2.0.0",
+    "rimraf": "^2.5.0"
+  }
+}
