+++ publish/publish.sh
@@ -1,0 +1,11 @@
+#!/bin/bash
+
+PACK_DIR=package;
+
+publish() {
+    cd $PACK_DIR
+    echo 'Publishing to npm...'
+    npm publish *.tgz
+}
+
+./pack.sh && publish
