+++ src/iqkeyboardmanager.android.ts
@@ -1,0 +1,4 @@
+import { ContentView } from "tns-core-modules/ui/content-view";
+
+export class PreviousNextView extends ContentView {
+}
