+++ src/iqkeyboardmanager.ios.ts
@@ -1,0 +1,7 @@
+import { ContentView } from "tns-core-modules/ui/content-view";
+
+export class PreviousNextView extends ContentView {
+  createNativeView() {
+    return IQPreviousNextView.new();
+  }
+}
