+++ src/package.json
@@ -1,0 +1,52 @@
+{
+  "name": "nativescript-iqkeyboardmanager",
+  "version": "1.3.0",
+  "description": "NativeScript wrapper of the popular IQKeyboardManager iOS library",
+  "main": "iqkeyboardmanager",
+  "typings": "index.d.ts",
+  "nativescript": {
+    "platforms": {
+      "ios": "3.0.0",
+      "android": "3.0.0"
+    }
+  },
+  "scripts": {
+    "build": "npm i && tsc",
+    "demo.ios": "cd ../demo && tns run ios --emulator --syncAllFiles",
+    "demo.ios.device": "cd ../demo && tns run ios --syncAllFiles",
+    "demo.android": "cd ../demo && tns run android --syncAllFiles",
+    "setup": "npm run build && cd ../demo && npm i",
+    "setupandinstall": "npm i && cd ../demo && npm i && cd ../src && npm run build && cd ../demo && tns plugin add ../src && cd ../src",
+    "tslint": "tslint --config '../tslint.json' '*.ts' --exclude '**/node_modules/**'",
+    "tslint.demo": "tslint --config '../tslint.json' '../demo/app/*.ts' --exclude '**/node_modules/**'",
+    "ci.tslint": "npm run tslint && npm run tslint.demo",
+    "prepublishOnly": "npm run build",
+    "development.setup": "npm run setup && npm link && cd ../demo && npm link nativescript-iqkeyboardmanager && cd ../src"
+  },
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/tjvantoll/nativescript-IQKeyboardManager.git"
+  },
+  "keywords": [
+    "NativeScript"
+  ],
+  "author": "TJ VanToll <tj.vantoll@gmail.com> (http://tjvantoll.com/)",
+  "license": "MIT",
+  "contributors": [
+    {
+      "name": "Eddy Verbruggen",
+      "email": "eddyverbruggen@gmail.com"
+    }
+  ],
+  "homepage": "https://github.com/tjvantoll/nativescript-IQKeyboardManager",
+  "bugs": {
+    "url": "https://github.com/tjvantoll/nativescript-IQKeyboardManager/issues"
+  },
+  "devDependencies": {
+    "nativescript-dev-typescript": "^0.6.0",
+    "tns-core-modules": "~3.4.0",
+    "tns-platform-declarations": "~3.4.0",
+    "typescript": "~2.6.0",
+    "tslint": "^5.0.0"
+  }
+}
