+++ src/references.d.ts
@@ -1,0 +1,2 @@
+/// <reference path="./node_modules/tns-platform-declarations/ios/ios.d.ts" />
+/// <reference path="./index.d.ts" />
